requirements.txt
gradio
accelerate
einops
pytorch-lightning
omegaconf
Pywavelets
transformers
open_clip_torch